Implement testing suite.
The test suite used in LoxLiaison was modified to be compatible with LoxMin as to accelerate testing. All the tests are kept the same, aside from 6 which had to be modified to accommodate for LoxMin's runtime errors being locational; that is, in order to properly parse a runtime error, LoxTestGenerator would also have to walk the Lox program itself to know where it was. Instead of doing this, the runtime errors were spelled out within the 6 tests to make life easier.